backend/requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-SQLAlchemy==3.0.3
SQLAlchemy==2.0.7
Werkzeug==2.2.3
python-dotenv==1.0.0
marshmallow==3.19.0
Pillow==9.4.0
gunicorn==20.1.0
psycopg2-binary==2.9.5
